+++ src/Aws/Glacier/Resources/glacier-2012-06-01.php
@@ -385,6 +385,34 @@ return array (
                 ),
             ),
         ),
+        'GetDataRetrievalPolicy' => array(
+            'httpMethod' => 'GET',
+            'uri' => '/{accountId}/policies/data-retrieval',
+            'class' => 'Guzzle\\Service\\Command\\OperationCommand',
+            'responseClass' => 'GetDataRetrievalPolicyOutput',
+            'responseType' => 'model',
+            'parameters' => array(
+                'accountId' => array(
+                    'required' => true,
+                    'type' => 'string',
+                    'location' => 'uri',
+                ),
+            ),
+            'errorResponses' => array(
+                array(
+                    'reason' => 'Returned if a parameter of the request is incorrectly specified.',
+                    'class' => 'InvalidParameterValueException',
+                ),
+                array(
+                    'reason' => 'Returned if a required header or parameter is missing from the request.',
+                    'class' => 'MissingParameterValueException',
+                ),
+                array(
+                    'reason' => 'Returned if the service cannot complete the request.',
+                    'class' => 'ServiceUnavailableException',
+                ),
+            ),
+        ),
         'GetJobOutput' => array(
             'httpMethod' => 'GET',
             'uri' => '/{accountId}/vaults/{vaultName}/jobs/{jobId}/output',
@@ -536,6 +564,10 @@ return array (
                 array(
                     'reason' => 'Returned if the specified resource, such as a vault, upload ID, or job ID, does not exist.',
                     'class' => 'ResourceNotFoundException',
+                ),
+                array(
+                    'reason' => 'Returned if a retrieval job would exceed the current data policy\'s retrieval rate limit. For more information about data retrieval policies,',
+                    'class' => 'PolicyEnforcedException',
                 ),
                 array(
                     'reason' => 'Returned if a parameter of the request is incorrectly specified.',
@@ -772,6 +804,55 @@ return array (
                     'reason' => 'Returned if the specified resource, such as a vault, upload ID, or job ID, does not exist.',
                     'class' => 'ResourceNotFoundException',
                 ),
+                array(
+                    'reason' => 'Returned if a parameter of the request is incorrectly specified.',
+                    'class' => 'InvalidParameterValueException',
+                ),
+                array(
+                    'reason' => 'Returned if a required header or parameter is missing from the request.',
+                    'class' => 'MissingParameterValueException',
+                ),
+                array(
+                    'reason' => 'Returned if the service cannot complete the request.',
+                    'class' => 'ServiceUnavailableException',
+                ),
+            ),
+        ),
+        'SetDataRetrievalPolicy' => array(
+            'httpMethod' => 'PUT',
+            'uri' => '/{accountId}/policies/data-retrieval',
+            'class' => 'Guzzle\\Service\\Command\\OperationCommand',
+            'responseClass' => 'EmptyOutput',
+            'responseType' => 'model',
+            'parameters' => array(
+                'accountId' => array(
+                    'required' => true,
+                    'type' => 'string',
+                    'location' => 'uri',
+                ),
+                'Policy' => array(
+                    'type' => 'object',
+                    'location' => 'json',
+                    'properties' => array(
+                        'Rules' => array(
+                            'type' => 'array',
+                            'items' => array(
+                                'name' => 'DataRetrievalRule',
+                                'type' => 'object',
+                                'properties' => array(
+                                    'Strategy' => array(
+                                        'type' => 'string',
+                                    ),
+                                    'BytesPerHour' => array(
+                                        'type' => 'numeric',
+                                    ),
+                                ),
+                            ),
+                        ),
+                    ),
+                ),
+            ),
+            'errorResponses' => array(
                 array(
                     'reason' => 'Returned if a parameter of the request is incorrectly specified.',
                     'class' => 'InvalidParameterValueException',
@@ -1118,6 +1199,33 @@ return array (
                 'SizeInBytes' => array(
                     'type' => 'numeric',
                     'location' => 'json',
+                ),
+            ),
+        ),
+        'GetDataRetrievalPolicyOutput' => array(
+            'type' => 'object',
+            'additionalProperties' => true,
+            'properties' => array(
+                'Policy' => array(
+                    'type' => 'object',
+                    'location' => 'json',
+                    'properties' => array(
+                        'Rules' => array(
+                            'type' => 'array',
+                            'items' => array(
+                                'name' => 'DataRetrievalRule',
+                                'type' => 'object',
+                                'properties' => array(
+                                    'Strategy' => array(
+                                        'type' => 'string',
+                                    ),
+                                    'BytesPerHour' => array(
+                                        'type' => 'numeric',
+                                    ),
+                                ),
+                            ),
+                        ),
+                    ),
                 ),
             ),
         ),