+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2.7.10 - Unreleased
 
 * Added support for user data to Amazon Elastic Transcoder.
+* Corrected the AWS Security Token Service endpoint. 
 
 ## 2.7.9 - 2014-12-08